99% finalizado

+++ src/main/java/com/example/real_state_consortium/models/Receipt.java
@@ -1,6 +1,5 @@
 package com.example.real_state_consortium.models;
 
-import com.example.real_state_consortium.utils.Data3;
 import java.time.LocalDate;
 import java.util.ArrayList;
 public class Receipt {
@@ -8,11 +7,11 @@ public class Receipt {
     private String nameBuyer;
     private ArrayList<Data3> elementsAdded;
     private Structure structure;
-    private Float priceSell;
+    private double priceSell;
     private LocalDate daySell;
     private boolean passed;
 
-    public Receipt(String nameAgent, String nameBuyer, ArrayList<Data3> elementsAdded, Structure structure , Float priceSell , LocalDate daySell , boolean passed) {
+    public Receipt(String nameAgent, String nameBuyer, ArrayList<Data3> elementsAdded, Structure structure , double priceSell , LocalDate daySell , boolean passed) {
         this.nameAgent = nameAgent;
         this.nameBuyer = nameBuyer;
         this.elementsAdded = elementsAdded;
@@ -40,68 +39,51 @@ public class Receipt {
         this.passed = passed;
     }
 
-
-
     public String getNameAgent() {
         return nameAgent;
     }
-
     public void setNameAgent(String nameAgent) {
         this.nameAgent = nameAgent;
     }
-
     public String getNameBuyer() {
         return nameBuyer;
     }
-
     public void setNameBuyer(String nameBuyer) {
         this.nameBuyer = nameBuyer;
     }
-
     public ArrayList<Data3> getElementsAdded() {
         return elementsAdded;
     }
-
     public void setElementsAdded(ArrayList<Data3> elementsAdded) {
         this.elementsAdded = elementsAdded;
     }
-
     public Structure getStructure() {
         return structure;
     }
-
     public void setStructure(Structure structure) {
         this.structure = structure;
     }
-
-    public Float getPriceSell() {
+    public Double getPriceSell() {
         return priceSell;
     }
-
-    public void setPriceSell(Float priceSell) {
+    public void setPriceSell(Double priceSell) {
         this.priceSell = priceSell;
     }
-
     public LocalDate getDaySell() {
         return daySell;
     }
-
     public void setDaySell(LocalDate daySell) {
         this.daySell = daySell;
     }
-
     public boolean isPassed() {
         return passed;
     }
-
     public void setPassed(boolean passed) {
         this.passed = passed;
     }
-
     @Override
     public String toString() {
-        return "Receipt{" +
-                "nameAgent='" + nameAgent + '\'' +
+        return "Receipt{" + "nameAgent='" + nameAgent + '\'' +
                 ", nameBuyer='" + nameBuyer + '\'' +
                 ", elementsAdded=" + elementsAdded +
                 ", structure=" + structure +
