+++ src/main/java/com/example/real_state_consortium/models/Structure.java
@@ -3,14 +3,14 @@ package com.example.real_state_consortium.models;
 public class Structure {
     private String typeStructure;
     private int floors;
-    private float front;
-    private float background;
+    private double front;
+    private double background;
     private int extractStructure;
-    private float squareMeters;
-    private float pricePerMeter2;
-    private float totalPriceOfStructure;
+    private double squareMeters;
+    private double pricePerMeter2;
+    private double totalPriceOfStructure;
     private boolean structureWasSell;
-    public Structure(String typeStructure, int floors, float front , float background , int extractStructure ,float squareMeters ,float pricePerMeter2,float totalPriceOfStructure) {
+    public Structure(String typeStructure, int floors, double front , double background , int extractStructure ,double squareMeters ,double pricePerMeter2,double totalPriceOfStructure) {
         this.typeStructure = typeStructure;
         this.floors = floors;
         this.front = front;
@@ -37,34 +37,34 @@ public class Structure {
     public void setFloors(int floors) {
         this.floors = floors;
     }
-    public float getFront() {
+    public double getFront() {
         return front;
     }
-    public void setFront(float front) {
+    public void setFront(double front) {
         this.front = front;
     }
-    public float getBackground() {
+    public double getBackground() {
         return background;
     }
-    public void setBackground(float background) {
+    public void setBackground(double background) {
         this.background = background;
     }
     public int getExtractStructure() {return extractStructure;}
     public void setExtractStructure(int extractStructure) {this.extractStructure = extractStructure;}
-    public float getSquareMeters() {
+    public double getSquareMeters() {
         return squareMeters;
     }
-    public void setSquareMeters(float squareMeters) {
+    public void setSquareMeters(double squareMeters) {
         this.squareMeters = squareMeters;
     }
-    public float getPricePerMeter2() {
+    public double getPricePerMeter2() {
         return pricePerMeter2;
     }
-    public void setPricePerMeter2(float pricePerMeter2) {
+    public void setPricePerMeter2(double pricePerMeter2) {
         this.pricePerMeter2 = pricePerMeter2;
     }
-    public float getTotalPriceOfStructure() {return totalPriceOfStructure;}
-    public void setTotalPriceOfStructure(float totalPriceOfStructure) {this.totalPriceOfStructure = totalPriceOfStructure;}
+    public double getTotalPriceOfStructure() {return totalPriceOfStructure;}
+    public void setTotalPriceOfStructure(double totalPriceOfStructure) {this.totalPriceOfStructure = totalPriceOfStructure;}
     public boolean getStructureWasSell() {return structureWasSell;}
     public void setStructureWasSell(boolean structureWasSell) {this.structureWasSell = structureWasSell;}
