+++ src/main/java/com/example/real_state_consortium/services/StructureService.java
@@ -3,10 +3,8 @@ package com.example.real_state_consortium.services;
 import com.example.real_state_consortium.controllers.ModelFactoryController;
 import com.example.real_state_consortium.models.Structure;
 
-import java.util.List;
-
 public interface StructureService {
-    void calculateStructure(ModelFactoryController mfc, float metersBackground, float meterFront, int floors, int extract, String stucture);
+    void calculateStructure(ModelFactoryController mfc, double metersBackground, double meterFront, int floors, int extract, String stucture);
     Structure getStructureClass();
     void setNullStructureClassObject();
